+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [[*next-version*]] - YYYY-MM-DD
+
+## [0.1-alpha8] - 2018-09-11
 ### Added
 - New mechanism for custom authorization using events.
 - Added handler to authorize certain WP apps by nonce.